refactor(reservation): extract shared UI components from page
- Create PriceBadge, PriceBreakdownCard, StepIndicator as standalone components
- Add barrel index.ts for reservation components
- Replace 3 inline component definitions in ReservationFlowPage with imports
- Zero behavior change, exact same props and store coupling

PR2/4 of refactor/reservation-flow chain
