+++ frontend/src/components/reservation/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Reservation Components - Barrel export
+ * Componentes de reserva
+ *
+ * @module components/reservation
+ */
+export { default as PriceBadge } from './price-badge';
+export { default as PriceBreakdownCard } from './price-breakdown-card';
+export { default as StepIndicator } from './step-indicator';
